requirements.txt
annotated-types==0.6.0
anyio==3.7.1
boto3==1.34.19
botocore==1.34.19
certifi==2023.11.17
ecdsa==0.18.0
fastapi==0.105.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jmespath==1.0.1
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
stripe==7.11.0
typing_extensions==4.9.0
urllib3==2.0.7
